light: add option to generate a lit file without modifying the bsp
Add command line option '-litonly' for a special case use where a
litfile can be generated to match an existing bsp exactly.  I.e. it must
generate coloured light data that exactly aligns with the existing light
data offsets in the bsp.

This of course has the limitation that a surface that was not lit in the
original bsp cannot be lit in the coloured light file since there will
not be any data allocation for it.  But it allows a .lit file that is
suitable for distribution to other people that have the original bsp
without requiring them to update the bsp file at the same time.