build: enable c99 + gnu extensions and fix silly cc-option bug
cc-option was defined after we were using it in the makefile, so the
cc option checking was not working and we never used the optional
parameters.

Pull in some of the cc option changes made to TyrQuake so we can do
the version check on gcc for -fno-unit-at-a-time (although I don't
have any gcc < 4.0 now).  Don't use -ffast-math since we haven't had
that enabled previously due to the cc-option bug and I suspect we
might see qbsp issues with that.

Make the osx build detect the supported architectures so I don't have
to manually do that when building a release.